+++ CHANGELOG.md
@@ -2,4 +2,6 @@
 
 ## [Unreleased]
 ### Added
-* franka_msgs package that contains common message, service and action type definitions
+* franka\_msgs package that contains common message, service and action type definitions
+* franka\_description package that contains all meshes and xacro files
+* franka\_gripper package that offers action and service interfaces to use the Franka Hand gripper